+++ spec/resource/organisation_http_spec.rb
@@ -20,31 +20,31 @@ describe 'Organisation' do
     end
 
     it '#address' do
-      expect(@p.address).to be_an_instance_of(Array) if @p.address
+      expect(@p.address).to be_an_instance_of(Array)
     end
 
     it '#name' do
-      expect(@p.name).to be_an_instance_of(String) if @p.name
+      expect(@p.name).to be_an_instance_of(String)
     end
 
     it '#organisations' do
-      expect(@p.organisations).to be_an_instance_of(Array) if @p.organisations
+      expect(@p.organisations).to be_an_instance_of(Array)
     end
 
     it '#parent' do
-      expect(@p.parent).to be_an_instance_of(Puree::Model::OrganisationHeader) if @p.parent
+      expect(@p.parent).to be_an_instance_of(Puree::Model::OrganisationHeader)
     end
 
     it '#phone_numbers' do
-      expect(@p.phone_numbers).to be_an_instance_of(Array) if @p.phone_numbers
+      expect(@p.phone_numbers).to be_an_instance_of(Array)
     end
 
     it '#type' do
-      expect(@p.type).to be_an_instance_of(String) if @p.type
+      expect(@p.type).to be_an_instance_of(String)
     end
 
     it '#urls' do
-      expect(@p.urls).to be_an_instance_of(Array) if @p.urls
+      expect(@p.urls).to be_an_instance_of(Array)
     end
 
   end