+++ spec/resource/project_http_spec.rb
@@ -20,55 +20,55 @@ describe 'Project' do
     end
 
     it '#acronym' do
-      expect(@p.acronym).to be_an_instance_of(String) if @p.acronym
+      expect(@p.acronym).to be_an_instance_of(String)
     end
 
     it '#description' do
-      expect(@p.description).to be_an_instance_of(String) if @p.description
+      expect(@p.description).to be_an_instance_of(String)
     end
 
     it '#organisations' do
-      expect(@p.organisations).to be_an_instance_of(Array) if @p.organisations
+      expect(@p.organisations).to be_an_instance_of(Array)
     end
 
     it '#owner' do
-      expect(@p.owner).to be_an_instance_of(Puree::Model::OrganisationHeader) if @p.owner
+      expect(@p.owner).to be_an_instance_of(Puree::Model::OrganisationHeader)
     end
 
     it '#persons_internal' do
-      expect(@p.persons_internal).to be_an_instance_of(Array) if @p.persons_internal if @p.persons_internal
+      expect(@p.persons_internal).to be_an_instance_of(Array)
     end
 
     it '#persons_external' do
-      expect(@p.persons_external).to be_an_instance_of(Array) if @p.persons_external if @p.persons_external
+      expect(@p.persons_external).to be_an_instance_of(Array)
     end
 
     it '#persons_other' do
-      expect(@p.persons_other).to be_an_instance_of(Array) if @p.persons_other if @p.persons_other
+      expect(@p.persons_other).to be_an_instance_of(Array)
     end
 
     it '#statuses' do
-      expect(@p.status).to be_an_instance_of(String) if @p.status
+      expect(@p.status).to be_an_instance_of(String)
     end
 
     it '#temporal_actual' do
-      expect(@p.temporal_actual).to be_an_instance_of(Puree::Model::TemporalRange) if @p.temporal_actual
+      expect(@p.temporal_actual).to be_an_instance_of(Puree::Model::TemporalRange)
     end
 
     it '#temporal_expected' do
-      expect(@p.temporal_expected).to be_an_instance_of(Puree::Model::TemporalRange) if @p.temporal_expected
+      expect(@p.temporal_expected).to be_an_instance_of(Puree::Model::TemporalRange)
     end
 
     it '#title' do
-      expect(@p.title).to be_an_instance_of(String) if @p.title
+      expect(@p.title).to be_an_instance_of(String)
     end
 
     it '#type' do
-      expect(@p.type).to be_an_instance_of(String) if @p.type
+      expect(@p.type).to be_an_instance_of(String)
     end
 
     it '#url' do
-      expect(@p.url).to be_an_instance_of(String) if @p.url
+      expect(@p.url).to be_an_instance_of(String)
     end
 
   end