+++ spec/resource/publication_http_spec.rb
@@ -18,59 +18,59 @@ describe 'Publication' do
     end
 
     it '#category' do
-      expect(@p.category).to be_an_instance_of(String) if @p.category
+      expect(@p.category).to be_an_instance_of(String)
     end
 
     it '#description' do
-      expect(@p.description).to be_an_instance_of(String) if @p.description
+      expect(@p.description).to be_an_instance_of(String)
     end
 
     it '#doi' do
-      expect(@p.doi).to be_an_instance_of(String) if @p.doi
+      expect(@p.doi).to be_an_instance_of(String)
     end
 
     it '#event' do
-      expect(@p.event).to be_an_instance_of(Puree::Model::EventHeader) if @p.event
+      expect(@p.event).to be_an_instance_of(Puree::Model::EventHeader)
     end
 
     it '#files' do
-      expect(@p.files).to be_an_instance_of(Array) if @p.files
+      expect(@p.files).to be_an_instance_of(Array)
     end
 
     it '#organisations' do
-      expect(@p.organisations).to be_an_instance_of(Array) if @p.organisations
+      expect(@p.organisations).to be_an_instance_of(Array)
     end
 
     it '#page' do
-      expect(@p.page).to be_an_instance_of(Fixnum) if @p.page
+      expect(@p.page).to be_an_instance_of(Fixnum)
     end
 
     it '#persons_internal' do
-      expect(@p.persons_internal).to be_an_instance_of(Array) if @p.persons_internal
+      expect(@p.persons_internal).to be_an_instance_of(Array)
     end
 
     it '#persons_external' do
-      expect(@p.persons_external).to be_an_instance_of(Array) if @p.persons_external
+      expect(@p.persons_external).to be_an_instance_of(Array)
     end
 
     it '#persons_other' do
-      expect(@p.persons_other).to be_an_instance_of(Array) if @p.persons_other
+      expect(@p.persons_other).to be_an_instance_of(Array)
     end
 
     it '#statuses' do
-      expect(@p.statuses).to be_an_instance_of(Array) if @p.statuses
+      expect(@p.statuses).to be_an_instance_of(Array)
     end
 
     it '#subtitle' do
-      expect(@p.subtitle).to be_an_instance_of(String) if @p.subtitle
+      expect(@p.subtitle).to be_an_instance_of(String)
     end
 
     it '#title' do
-      expect(@p.title).to be_an_instance_of(String) if @p.title
+      expect(@p.title).to be_an_instance_of(String)
     end
 
     it '#type' do
-      expect(@p.type).to be_an_instance_of(String) if @p.type
+      expect(@p.type).to be_an_instance_of(String)
     end
 
   end