+++ CHANGELOG.md
@@ -5,11 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 - Metadata: activity?, clipping?, externalPerson?
 - Factory to make resource objects?
-- Make ALL dates ISO 8601 YYYY-MM-DD, rather than mirror varying formats from Pure?
 
 ## 1.0.0 - 2017-02-22
 ### Added
-- Dataset - legal_conditions.
+- Dataset - legal conditions.
 
 ### Changed
 - Simplified basic auth.